Working on part 2
